+++ ADsurf/_cps/_surf96_vectorAll.py
@@ -323,7 +323,7 @@ def dltar4_vector(pvs_lists, t_lists, d, alpha, beta, rho, llw=1):
 
     # dunkin's matrix [station,pvs_points,vs,5]
     e = torch.zeros((wvno.shape[0],wvno.shape[1],alpha.shape[1],5))
-    omega = torch.max(omega, torch.tensor(1.0e-4))
+    omega = torch.max(omega, torch.tensor(1.0e-4).to(torch.float32))
     wvno2 = wvno * wvno
     xka = omega / alpha[:,[-1]] # 切片之后不要丢失维度
     xkb = omega / beta[:,[-1]]
